+++ Assets/_Scripts/Units/UI/Menus/MainMenu.cs
@@ -1,0 +1,15 @@
+﻿using TMPro;
+using UnityEngine;
+
+namespace Metroidvania.UI.Menus
+{
+    public class MainMenu : MonoBehaviour
+    {
+        [SerializeField] private TextMeshProUGUI versionText;
+        
+        private void Awake()
+        {
+            versionText.text = Application.version;
+        }
+    }
+}