+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ﻿# 1.2.6-a.2
 ## Release Highlights
 - Upgrade Unity to 2021.3.3f1
-- Add Unity AddressablesSystem
+- Add Unity Addressables System
 - Add Entities (enemies) system
 - Fix code comments and typos
 - Add singleton ScriptableObject behaviour called [ScriptableSingleton\<T>](Assets/Project/Runtime/Utility/ScriptableSingleton.cs)
@@ -17,3 +17,19 @@
 - Add Editor class to handle project validation
 - Add ScriptableObject to handle a entity instance
 - Add ScriptableObject to handle all entities
+
+# 1.3.0-a.1
+# Release Highlightstings
+- Add Audio system
+- Add menus (main title, pause and options (settings management))
+- Add Unity Localization System
+- Add TextMesh Pro font
+- Update project structure
+
+## Features
+- Add UI Fields
+- Change sorting layers
+- Add texture importer
+- Add RangedFloat/Int fields with custom editor drawer
+- Add UI sounds
+- Removed default TextMesh Pro assets